Upload 20211230

+++ updater/create_databases/upload_new.py
@@ -6,7 +6,7 @@ import pandas as pd
 from sqlalchemy import create_engine
 
 from QA.create_databases.UploadTest import UploadTest
-from lib.configuration import get_connection_string, get_required_tables
+from lib.configuration import get_connection_string, get_required_tables, get_current_config
 
 
 def upload_table(table_name, filepath, connection_string, version_indicator):
@@ -31,33 +31,39 @@ def consolidate_cpc_classes(connection_string):
     for table_name in ['mainclass', 'subclass']:
         engine = create_engine(connection_string)
         insert_statement = "INSERT IGNORE INTO {table_name} (id) SELECT id from temp_{table_name};".format(
-                table_name=table_name)
+            table_name=table_name)
         engine.execute(insert_statement)
         engine.dispose()
 
 
-def setup_database(update_config):
+def setup_database(update_config, drop=True):
     required_tables = get_required_tables(update_config)
     print("Required tables are {tlist}".format(tlist=", ".join(required_tables)))
-    connection_string = get_connection_string(update_config, "RAW_DB")
+    connection_string = get_connection_string(update_config, database="PROD_DB")
     engine = create_engine(connection_string)
-    raw_database = update_config["PATENTSVIEW_DATABASES"]["RAW_DB"]
+    raw_database = update_config["PATENTSVIEW_DATABASES"]["PROD_DB"]
     temp_upload_database = update_config["PATENTSVIEW_DATABASES"]["TEMP_UPLOAD_DB"]
+    if drop:
+        engine.execute("""
+            DROP DATABASE if exists {temp_upload_database}
+        """.format(temp_upload_database=temp_upload_database))
     engine.execute("""
-DROP DATABASE if exists {temp_upload_database}
-    """.format(temp_upload_database=temp_upload_database))
-    engine.execute("""
-create database if not exists {temp_upload_database} default character set=utf8mb4 default collate=utf8mb4_unicode_ci
-            """.format(
-            temp_upload_database=temp_upload_database))
+            create database if not exists {temp_upload_database} default character set=utf8mb4
+             default collate=utf8mb4_unicode_ci
+        """.format(
+        temp_upload_database=temp_upload_database))
     for table in required_tables:
         print("Creating Table : {tbl}".format(tbl=table))
         con = engine.connect()
-        con.execute("drop table if exists {0}.{1}".format(temp_upload_database, table))
-        con.execute(
+        if drop:
+            con.execute("drop table if exists {0}.{1}".format(temp_upload_database, table))
+        if table in ['inventor', 'assignee_disambiguation_mapping', 'inventor_disambiguation_mapping', 'assignee']:
+            con.execute("create table if not exists {0}.{2} like {1}.{2}".format(temp_upload_database, 'upload_20211130', table))
+        else:
+            con.execute(
                 "create table if not exists {0}.{2} like {1}.{2}".format(temp_upload_database, raw_database, table))
-        con.close()
-    engine.dispose()
+            con.close()
+        engine.dispose()
 
 
 def generate_timestamp_uploads(update_config):
@@ -91,12 +97,18 @@ def upload_current_data(**kwargs):
     begin_upload(update_config=config)
 
 
-def post_upload(**kwargs):
+def post_upload_granted(**kwargs):
     from lib.configuration import get_current_config
     config = get_current_config('granted_patent', **kwargs)
     qc = UploadTest(config)
     qc.runTests()
 
+
+def post_upload_pgpubs(**kwargs):
+    from lib.configuration import get_current_config
+    config = get_current_config('pgpubs', **kwargs)
+    qc = UploadTest(config)
+    qc.runTests()
 
 if __name__ == '__main__':
     begin_database_setup(**{
@@ -108,3 +120,9 @@ if __name__ == '__main__':
     # post_upload(**{
     #         "execution_date": datetime.date(2020, 12, 1)
     #         })
+    config = get_current_config('granted_patent', **{
+        "execution_date": datetime.date(2020, 12, 14)
+    })
+    setup_database(config, **{
+            "execution_date": datetime.date(2020, 12, 14)
+            })
